tests user compaction requested using FlakyAmple
Through analyzing code coverage data it was discovered
that the conditional mutations related to user compaction
requested were not being exercised in tests.  This commit
refactors an existing test so that it can run with and
without FlakyAmple which will excercise that conditional
mutation.
